Add UIApplication.shared.requestPermission
